+++ Core/Entity/Item/Containerable/Effect/ItemMap.cs
@@ -16,7 +16,7 @@ class ItemMap : ItemContainerable
 		Size = new Vector2(50, 50);
 		Name = "Gymvod plánek";
 		Description = "S ním se neztratíš. \n(minimapu otevřeš M)\n";
-		Stats = new Dictionary<StatTypes, int>() { };
+		Stats = new Dictionary<StatTypes, float>() { };
 		Effects = new List<EffectTypes>() { EffectTypes.MAP_REVEAL };
 		Sprite = TextureManager.GetTexture("gymvodMap");
 		ItemType = ItemTypes.EFFECT;
